Add functionality to support meta values in mod config file.
I ran into this issue when trying to use Metallurgy 3. The config file
in this mod specifies the objects with meta values:
S:"Angmallen Block":910:6

Previously, this would throw an IllegalNumberFormatException. This now
properly parses the object id from the meta value. The meta value added
in the mod config file will be added to the ModAspectEntry, regardless
of what meta values have been specified in the aspect config file.